part7_thuesday

+++ src/main/java/com/example/insuranceapplication/controller/paymentController/PaymentController.java
@@ -1,13 +1,10 @@
 package com.example.insuranceapplication.controller.paymentController;
 
-import com.example.insuranceapplication.entity.ClientPassword;
 import com.example.insuranceapplication.entity.payment.Payment;
-import com.example.insuranceapplication.entity.updateDto.ClientPasswordUpdateDto;
 import com.example.insuranceapplication.entity.updateDto.PaymentUpdateDto;
 import com.example.insuranceapplication.service.database.PaymentDatabaseService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-//import org.springframework.security.core.parameters.P;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -20,7 +17,7 @@ public class PaymentController {
     private final PaymentDatabaseService paymentDatabaseService;
 
     @GetMapping(value = "/payment/find/amount/{amount}")
-    public ResponseEntity<List<Payment>> getPaymentsByAmount(@PathVariable(name = "amount") Integer amount) {
+    public ResponseEntity<List<Payment>> getPaymentsByAmount(@PathVariable(name = "amount") Double amount) {
         List<Payment> payments = paymentDatabaseService.getPaymentsByAmount(amount);
         return createResponseEntity(payments);
     }
